pyproject.toml
[project]
name = "czii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.0",
    "connected-components-3d>=3.22.0",
    "copick>=0.7.0",
    "copick-utils",
    "gdown>=5.2.0",
    "imantics>=0.1.12",
    "jupyter>=1.1.1",
    "kaggle>=1.6.17",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.9.3",
    "monai-weekly[mlflow]>=1.5.dev2449",
    "monai[all]>=1.4.0",
    "numpy==1.26.4",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "pytorch-lightning>=2.5.0.post0",
    "pyzmq>=26.2.0",
    "sam-2",
    "schedulefree>=1.4",
    "scikit-image==0.22.0",
    "scikit-learn==1.3.2",
    "timm>=1.0.12",
    "torchinfo>=1.8.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.53",
    "volumentations-3d>=1.0.4",
    "wandb>=0.19.5",
    "zarr>=2.18.3",
]

[tool.uv.sources]
copick-utils = { git = "https://github.com/copick/copick-utils.git" }
sam-2 = { git = "https://github.com/facebookresearch/segment-anything-2/" }

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
